+++ src/ecoshard/__main__.py
@@ -4,11 +4,16 @@ import configparser
 import glob
 import hashlib
 import logging
+import multiprocessing
 import os
 import sys
 import threading
 
+from ecoshard import taskgraph
+from osgeo import gdal
 import ecoshard
+
+gdal.SetCacheMax(2**26)
 
 LOGGER = logging.getLogger(__name__)
 
@@ -19,9 +24,11 @@ logging.basicConfig(
         '%(name)s [%(funcName)s:%(lineno)d] %(message)s'),
     stream=sys.stdout)
 logging.getLogger('ecoshard').setLevel(logging.DEBUG)
+logging.getLogger('ecoshard.taskgraph').setLevel(logging.INFO)
 LOGGER = logging.getLogger(__name__)
 
 CONFIG_PATH = os.path.expanduser(os.path.join('~', 'ecoshard.ini'))
+HASH_LEN_DEFAULT = 6
 
 
 def main():
@@ -122,6 +129,11 @@ def main():
             'Hash the file and and rename/copy depending on if --rename is '
             'set.'))
     process_subparser.add_argument(
+        '--hash_length', type=int, default=HASH_LEN_DEFAULT, help=(
+            'Limit the length of the hash to this many characters. '
+            f'Default is {HASH_LEN_DEFAULT}.'))
+
+    process_subparser.add_argument(
         '--force', action='store_true', help=(
             'force an ecoshard hash if the filename looks like an ecoshard. '
             'The new hash will be appended to the filename.'))
@@ -199,58 +211,38 @@ def main():
             args.datetime, args.asset_id, args.catalog_list)
         return 0
 
-    for glob_pattern in args.filepath:
-        for file_path in glob.glob(glob_pattern):
-            working_file_path = file_path
-            LOGGER.info('processing %s', file_path)
+    file_list = [
+        file_path
+        for glob_pattern in args.filepath
+        for file_path in glob.glob(glob_pattern)]
 
-            if args.reduce_factor:
-                method = args.reduce_factor[1]
-                valid_methods = ["max", "min", "sum", "average", "mode"]
-                if method not in valid_methods:
-                    LOGGER.error(
-                        '--reduce_method must be one of %s' % valid_methods)
-                    sys.exit(-1)
-                ecoshard.convolve_layer(
-                    file_path, int(args.reduce_factor[0]),
-                    args.reduce_factor[1],
-                    args.reduce_factor[2])
-                continue
+    n_workers = min(multiprocessing.cpu_count(), len(file_list))
+    taskgraph_dir = '_ecoshard_taskgraph_dir_ok_to_delete'
+    os.makedirs(taskgraph_dir, exist_ok=True)
+    task_graph = taskgraph.TaskGraph(taskgraph_dir, n_workers, 15.0)
 
-            if args.compress:
-                prefix, suffix = os.path.splitext(file_path)
-                compressed_filename = '%s_compressed%s' % (prefix, suffix)
-                ecoshard.compress_raster(
-                    file_path, compressed_filename,
-                    compression_algorithm='DEFLATE')
-                working_file_path = compressed_filename
+    result_list = []
+    error_messages = []
+    for file_path in file_list:
+        result = task_graph.add_task(
+            func=ecoshard.process_worker,
+            args=(file_path, args,),
+            task_name=f'processing {file_path}')
+        result_list.append(result)
 
-            if args.buildoverviews:
-                overview_token_path = '%s.OVERVIEWCOMPLETE' % (
-                    working_file_path)
-                ecoshard.build_overviews(
-                    working_file_path, target_token_path=overview_token_path,
-                    interpolation_method=args.interpolation_method)
+    for result in result_list:
+        # ensure no bad result
+        try:
+            result.join()
+        except Exception as e:
+            error_messages.append(str(e))
 
-            if args.validate:
-                try:
-                    is_valid = ecoshard.validate(working_file_path)
-                    if is_valid:
-                        LOGGER.info('VALID ECOSHARD: %s', working_file_path)
-                    else:
-                        LOGGER.error(
-                            'got a False, but no ValueError on validate? '
-                            'that is not impobipible?')
-                except ValueError:
-                    LOGGER.error('INVALID ECOSHARD: %s', working_file_path)
-                    return_code = -1
-            elif args.hash_file:
-                hash_token_path = '%s.ECOSHARDCOMPLETE' % (
-                    working_file_path)
-                ecoshard.hash_file(
-                    working_file_path, target_token_path=hash_token_path,
-                    rename=args.rename, hash_algorithm=args.hashalg,
-                    force=args.force)
+    for message in error_messages:
+        LOGGER.error(message)
+        return_code = -1
+
+    task_graph.join()
+    task_graph.close()
     return return_code