+++ src/ecoshard/geoprocessing/routing/__init__.py
@@ -13,8 +13,6 @@ from ecoshard.geoprocessing.routing.routing import (
     calculate_subwatershed_boundary,
     detect_lowest_drain_and_sink,
     )
-from ecoshard.geoprocessing.routing.watershed import (
-    delineate_watersheds_d8)
 
 __all__ = (
     'fill_pits',
@@ -26,7 +24,6 @@ __all__ = (
     'distance_to_channel_d8',
     'distance_to_channel_mfd',
     'extract_streams_mfd',
-    'delineate_watersheds_d8',
     'detect_outlets',
     'extract_strahler_streams_d8',
     'calculate_subwatershed_boundary',